Move pages to routes
