+++ frontend/src/routes/MyApplicationsPage.tsx
